renamed file to types.ts
